KKUCAssistant/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.10.6

# LangChain dependencies (compatible versions)
langchain-core>=0.2.43,<0.4.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph==0.2.59
langsmith>=0.1.0

# LangServe for API
langserve[all]>=0.0.30

# Web scraping tool
firecrawl-py==1.10.0

# Additional required dependencies
httpx>=0.25.0

# Google Calendar API dependencies
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
pytz==2025.2
